requirements.txt
assemblyai==0.19.0
streamlit==1.31.0
plotly==5.18.0
pandas==2.2.0
python-dotenv==1.0.0
